chilli: Fix a build error with --enable-json
This fixes a build error that occurs when --enable-json config flag is
set and libjson-c is not available. This adds a check in configure.ac
for availability of libjson-c library.

Also libjson-c wasnt present in the list of linked libraries for chilli
binary. Hence a linker error complaining about missing json_object_xxx
functions used to occur. To fix this a new m4 macro named LIBJSON is
added to configure.ac that expands to -ljson-c and is added to list of
libraries to be linked to chill in src/Makefile.am.

Signed-off-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>