Ability to configure the Syslog Identity
This patch is about Syslog. When applied, coova-chilli will use the
filename as Syslog identity.
This is an enhancement, to be able to launch 2 instances of the daemon
with different Syslog identities.
For issue #275, we could use a symbolic links to the real filename and
execute coova-chilli using those links.
This is better than adding a new option to the already long list !

Signed-off-by: Baligh GUESMI <gasmibal@gmail.com>